Add realtime update for the Gauge list fragment
This approach works by introducing RealtimeTrafficService,
which manages the Pusher previously associated only with the
AirTraffic activity. Interested activities and fragments
bind to the service, and register themselves as listeners for
Hit events.

This has the advantage that the same pusher can be initialised
and then shared amongst multiple activities - so start up of
the AirTraffic activity can now be immediate, rather than the
10-second-ish delay we currently have.

Only the view count is incremented at the moment, but the
Gauges web app has code which shows us how we can use the
'uniques' data to increment the 'people' counter too.